feat: add VTEX account and workspace management functionalities
- Introduced new module for handling VTEX account operations including switching accounts and retrieving account information.
- Implemented caching mechanism for workspaces to improve performance and reduce redundant API calls.
- Added status bar integration to display current VTEX account and workspace, with options to switch accounts and workspaces.
- Created utility functions for reading and parsing VTEX configuration files (session.json and workspace.json).
- Enhanced error handling and logging throughout the modules for better debugging and user feedback.
- Implemented command execution in terminal with options for silent execution and output capture.
