+++ src/shared/index.ts
@@ -1,4 +1,5 @@
-export * from "./helpers";
-export * from "./models";
-export * from "./utils";
-export * from "./watchers";
+export * from './helpers';
+export * from './models';
+export * from './utils';
+export * from './vtex';
+export * from './watchers';
