+++ src/shared/vtex/index.ts
@@ -1,0 +1,5 @@
+export * from './account';
+export * from './cache';
+export * from './statusBar';
+export * from './terminal';
+export * from './workspace';
